Implemented vectorization for ``total_cost`` computation.
The objective is to avoid looping through each individual example to compute the cost.
In order to do this, we define a single matrix containing all the inputs for training/validation data. However, the ``accuracy`` function needs the non-vectorized results.
Now, at the begining of ``SGD`` I save the data in both formats. Probably inefficient use of memory, but it should speed things up.